Fix rendering on GTA V version 1.27
We want to support the 'speedrunner patch' version of the game.

Unfortunately, v1.27 suffers from a weird error where drawing any 3D boxes completely messes up the positioning of the line drawing, so we solve this by doing two loops through the stunt jumps, drawing the solid boxes after all wireframes have already been drawn.

This has a small disadvantage of 3D boxes rendering *over* the
wireframes, but this is the fastest way to fix this bug I could find.